+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.3.2
+
+### Patch Changes
+
+-   7d61e58: Fixed writing collections with string values
+
+## 0.3.1
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.28.8
+-   @sap-ux/cds-odata-annotation-converter@0.3.9
+
 ## 0.3.0
 
 ### Minor Changes